Store timestamps as longs, since ints are 16 bits on Arduino.
... This limits the working time to about 32 seconds :D.

+++ src/AbstractLightController.h
@@ -14,10 +14,10 @@ public:
   virtual void setOn() = 0;
   virtual void setOff() = 0;
   virtual void setAnimating() = 0;
-  virtual void gradualOn(int transitionUntilMs) = 0;
-  virtual void gradualOff(int transitionUntilMs) = 0;
+  virtual void gradualOn(unsigned long transitionUntilMs) = 0;
+  virtual void gradualOff(unsigned long transitionUntilMs) = 0;
 
-  virtual void clockTick(int currentTimeMs) = 0;
+  virtual void clockTick(unsigned long currentTimeMs) = 0;
 };
 
 #endif
